Moved some functions to utils.py
Also worked on some PEP8 issues.
